+++ lib/ecto/adapters/sqlite3/connection.ex
@@ -1369,10 +1369,14 @@ defmodule Ecto.Adapters.SQLite3.Connection do
     [quote_name(name)]
   end
 
+  @datetime_type Application.compile_env(:ecto_sqlite3, :datetime_type, :iso8601)
+
   defp expr({:datetime_add, _, [datetime, count, interval]}, sources, query) do
+    format = Ecto.Adapters.SQLite3.Codec.datetime_format(@datetime_type)
+
     [
       "CAST (",
-      "strftime('%Y-%m-%d %H:%M:%f000Z'",
+      "strftime('#{format}'",
       ",",
       expr(datetime, sources, query),
       ",",